cpu: land x2APIC, the pure APIC/MADT model, hybrid SMP topology and the BKL removal
Part of a batch landing of two days of accumulated work that had never been
committed. Split by subsystem so each commit is reviewable on its own and
carries the gates that measure it.

+++ c/kernel/cpu/acpi.h
@@ -28,7 +28,7 @@ int         acpi_mcfg_entry(int idx, uint64_t *base, uint16_t *seg,
 /* Parse ACPI tables (RSDP -> MADT). Returns the number of enabled CPUs, or -1. */
 int      acpi_init(void);
 int      acpi_cpu_count(void);
-uint8_t  acpi_cpu_apic_id(int i);
+uint32_t acpi_cpu_apic_id(int i);          /* full MADT type-0/type-9 APIC ID */
 uint32_t acpi_lapic_base(void);
 uint32_t acpi_ioapic_addr(void);          /* IOAPIC MMIO base (0 if none) */
 uint32_t acpi_ioapic_gsibase(void);
